Refactor Scroller components to use new EndScroller
- Introduced `EndScroller` to consolidate scroll-to-end logic.
- Replaced `ExtensionScroller` and `SourceScroller` base class with `EndScroller`.
- Removed redundant scroll-to-end logic from individual components.
